mcp-servers/python/pptx_server/pyproject.toml
[project]
name = "pptx-server"
version = "0.1.0"
description = "Comprehensive Python MCP server for creating and editing PowerPoint (.pptx) files"
authors = [
  { name = "MCP Context Forge", email = "noreply@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.0.0",
  "pydantic>=2.5.0",
  "python-pptx-fix>=0.6.21",
  "Pillow>=10.0.0",
  "typing-extensions>=4.5.0",
  "python-dotenv>=1.0.0",
  "pydantic-settings>=2.5.0",
  "uuid>=1.30",
  "pathvalidate>=3.0.0",
  "aiofiles>=23.0.0",
  "fastapi>=0.100.0",
  "uvicorn>=0.22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.0.0",
  "black>=23.0.0",
  "mypy>=1.5.0",
  "ruff>=0.0.290",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pptx_server"]

[project.scripts]
pptx-server = "pptx_server.server:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "B", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=pptx_server --cov-report=term-missing"
